Add: 공동 컴포넌트 수정 및 팀원 배포 준비 완료
