+++ src/Components/Nav/Category2/Category2.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 
 import '../../../Styles/common.scss';
 import '../../../Styles/reset.scss';
-import '../../Main/Component/Category2.scss';
+import '../Category2/Category2.scss';
 
 class Category2 extends Component {
   render() {
