--- src/Pages/Main/Component/Footer.scss
@@ -1,61 +1,0 @@
-footer {
-  margin-top: 100px;
-  width: 100%;
-  color: white;
-  background-color: black;
-
-  .footerTop {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-    height: 160px;
-    background-color: lightgray;
-  }
-
-  .footer {
-    display: flex;
-    justify-content: space-around;
-    align-items: center;
-    padding: 100px 500px;
-    width: 100%;
-    height: 450px;
-
-    .footerLeft {
-      width: 30%;
-      font-size: 35px;
-    }
-
-    .footerRight {
-      width: 70%;
-
-      form {
-        display: flex;
-        height: 45px;
-
-        input {
-          width: 400px;
-          border-radius: 0;
-          height: 45px;
-          font-family: 'Gothic A1', sans-serif;
-        }
-
-        button {
-          width: 100px;
-          border: 1px white solid;
-          background-color: black;
-          color: white;
-          font-family: 'Gothic A1', sans-serif;
-        }
-      }
-
-      p {
-        margin: 10px 0;
-        width: 500px;
-        font-size: 15px;
-        font-weight: bolder;
-        font-family: 'Gothic A1', sans-serif;
-      }
-    }
-  }
-}
