--- src/Pages/Main/Component/MainAll.scss
@@ -1,50 +1,0 @@
-main {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-  height: auto;
-  // background-color: royalblue;
-
-  section:nth-child(3) {
-    display: flex;
-    justify-content: center;
-    width: 100%;
-
-    .main3Wrapper {
-      display: flex;
-      margin: 0 400px;
-      width: 800px;
-      height: 500px;
-
-      div:nth-child(1) {
-        margin-right: 20px;
-        width: 50%;
-
-        h3 {
-          margin: 100px 0 20px 0;
-          font-size: 40px;
-          font-family: 'Gothic A1', sans-serif;
-        }
-
-        p {
-          font-size: 15px;
-          font-family: 'Gothic A1', sans-serif;
-          font-weight: bolder;
-        }
-      }
-
-      div:nth-child(2) {
-        width: 50%;
-        height: 100%;
-        overflow: hidden;
-
-        img {
-          width: 100%;
-          height: 100%;
-          object-fit: cover;
-        }
-      }
-    }
-  }
-}
