+++ src/Pages/Main/Component/MainAll/MainAll.scss
@@ -1,0 +1,7 @@
+main {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+  height: auto;
+}
