+++ src/Pages/Main/Component/MainAll/MainSection1/MainSection1.scss
