+++ src/Pages/Main/Component/MainAll/MainSection2/MainSection2.scss
