--- src/Pages/Main/Component/Nav.scss
@@ -1,52 +1,0 @@
-nav {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  position: fixed;
-  top: 0;
-  width: 100%;
-  height: 100px;
-  background-color: black;
-  font-size: 20px;
-
-  h1 {
-    margin-right: 150px;
-    color: white;
-    letter-spacing: 2px;
-  }
-
-  ul {
-    display: flex;
-    justify-content: space-around;
-    align-items: center;
-    // width: 100%;
-    height: 100%;
-    color: white;
-
-    li {
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      // width: 100px;
-      margin-right: 60px;
-      height: 100%;
-      font-size: 16px;
-      font-weight: bold;
-      font-family: 'Gothic A1', sans-serif;
-    }
-
-    li:nth-child(6) {
-      color: yellow;
-    }
-  }
-
-  .navIcon {
-    li {
-      margin: 0 20px;
-
-      i {
-        font-size: 20px;
-      }
-    }
-  }
-}
